+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 ### Changed
+
+- Behavior of `oneOf` in the rust simulator now matches the typescript one (#1773)
+
 ### Deprecated
 ### Removed
 ### Fixed