+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 ### Removed
 ### Fixed
+
+- Fixed copying and pasting multi-line expressions from the REPL (#1776)
+
 ### Security
 
 ## v0.28.0 -- 2025-09-16