+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Behavior of `oneOf` in the rust simulator now matches the typescript one (#1773)
 - The REPL now prints file embedded expressions when executing them (#1774)
 
 ### Deprecated